Use an empty designator in .IP0
This seems to prevent groff -man -Thtml from interpreting a .IP without
a designator as the body of a preceding .TP without a body.

The author of this commit has waived all copyright and
related or neighboring rights to the extent permitted by
law as described by the CC0 1.0 Universal Public Domain
Dedication. You should have received a copy of the full
dedication along with this commit, typically as a file
named <CC0-1.0.txt>. If not, it may be available at
<https://creativecommons.org/publicdomain/zero/1.0/>.