Improve the compatibility of gatbps_check_macros
Commit ccd138672249a105e690beee24643fb6024a48b8 changed the formatting
of gatbps_check_macros to put the macro calls on a single line instead
of putting each on its own line. This was fine, but the two formats do
not mix well, so if you happened to have certain files with different
versions, then the construction of gatbps_check_macros could be wrong,
and the normal way to call it could produce a broken configure script.

This commit changes the formatting back to putting each macro call on
its own line. Having one long line is a bad idea anyway.

The author of this commit has waived all copyright and
related or neighboring rights to the extent permitted by
law as described by the CC0 1.0 Universal Public Domain
Dedication. You should have received a copy of the full
dedication along with this commit, typically as a file
named <CC0-1.0.txt>. If not, it may be available at
<https://creativecommons.org/publicdomain/zero/1.0/>.